[level 1] Title: 두 정수 사이의 합, Time: 12.20 ms, Memory: 60.5 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.kt
@@ -3,8 +3,8 @@ import kotlin.math.min
 
 class Solution {
     fun solution(a: Int, b: Int): Long {
-        val min = min(a, b).toLong()
-        val max = max(a, b).toLong()
+        val min = min(a, b)
+        val max = max(a, b)
         
         return (min..max).sum().toLong()
     }
